online_server/requirements.txt
fastapi
uvicorn
mysql-connector-python
sqlalchemy
# aiomysql
python-decouple
packaging
pymysql
